Add checks to make sure instrument is in the specified lab, that operator is permitted to use the instrument, add compressibility to buffer table, minor fixes

+++ sql/us3_proj_procs.sql
@@ -14,12 +14,12 @@ DROP FUNCTION IF EXISTS verify_project_permission$$
 CREATE FUNCTION verify_project_permission( p_guid      CHAR(36),
                                            p_password  VARCHAR(80),
                                            p_projectID INT )
-  RETURNS TINYINT
+  RETURNS INT
   READS SQL DATA
 
 BEGIN
   DECLARE count_projects INT;
-  DECLARE status         TINYINT;
+  DECLARE status         INT;
 
   CALL config();
   SET status   = @ERROR;
